modal for inline editing, 'completionArray' deleted
If the user wants to edit any of the to-do tasks that are already on the to-do list UI, then the user can press on either the task or the due date, which will then bring a modal that has two inputs with the tasks and the due dates as its values. If the user made any edits, then the user can press confirm. If the user wants to disregard any changes/edits, then the user can press cancel. Still working on the styling for the modals.
Deleted the 'completionArray'. The rendering of the DOM will be based upon the key name of 'completed' in each item in the 'tasksArray'. If 'completed: false', then the 'task value' of the object will be rendered onto the to-do list of the DOM. If 'completed: true', then the task of that object will be rendered onto the completion list of the DOM.