list styling updated, modalForClears reduced
Added an overflow property for the task items on each list. If the user has a long task length, the overflow property will be able to contain the words in their respective div. 
Reduce the modalForClears component down to one if statement and one else statement.
Added a form validation for the date input (date cannot be more than 10 characters long) and for the task input (input cannot just be blank spaces).